requirements.txt
Django>=5.0,<5.1
djangorestframework>=3.14.0
Pillow>=10.0.0
django-filter>=23.0
django-crispy-forms>=2.0
crispy-bootstrap4>=2.0
python-dateutil>=2.8.2
